requirements.txt
requests>=2.32
Pillow>=10.4
pydantic>=2.8
python-dotenv>=1.0
tenacity>=8.5
APScheduler>=3.10
googlemaps>=4.10
openai>=1.40
beautifulsoup4>=4.12
requests-oauthlib>=2.0
PyYAML>=6.0
loguru>=0.7
psycopg[binary]>=3.2
pytest>=8.2
pytest-mock>=3.14
